refactor: centralize time utilities
